refactor: server unit tests
